transcripter/requirements.txt
fastapi
uvicorn
python-multipart
requests
google-generativeai
streamlit
uv
